+++ src/uarch-record-state-access.h
@@ -51,7 +51,7 @@ class uarch_record_state_access : public i_uarch_state_access<uarch_record_state
     /// \param length Length of physical memory region.
     /// \returns Corresponding entry if found, or a sentinel entry
     /// for an empty range.
-    pma_entry &find_memory_pma_entry(uint64_t paddr, size_t length) {
+    pma_entry &find_memory_pma_entry(uint64_t paddr, uint64_t length) {
         // First, search microarchitecture's private PMA entries
         if (m_us.ram.contains(paddr, length)) {
             return m_us.ram;
